add user config for line_no and font size
these attributes can't be changes after initial conifiguration in the display.

Another attribute that can't be changed is the interface width, however
that's an implementation detail, no need for the user to change it.

Even though there are these configuration parameters, the 1-line mode
isn't handled by the driver beside the configuration. (E.g. it will be happy
trying to switch line in 1-line mode also.)